backend/requirements.txt
requests
Gunicorn
Flask
flask-sock
Flask-SQLAlchemy
Flask-Login
Flask-Cors
psycopg2-binary
python_dotenv
